size_t, all tests ran

+++ OndselMbD/ASMTAnimationParameters.h
@@ -17,9 +17,9 @@ namespace MbD {
     public:
         static std::shared_ptr<ASMTAnimationParameters> With();
         void parseASMT(std::vector<std::string>& lines) override;
-        void storeOnLevel(std::ofstream& os, int level) override;
+        void storeOnLevel(std::ofstream& os, size_t level) override;
 
-		int nframe = 1000000, icurrent = 1, istart = 1, iend = 1000000, framesPerSecond = 30;
+		size_t nframe = 1000000, icurrent = 1, istart = 1, iend = 1000000, framesPerSecond = 30;
         bool isForward = true;
