+++ OndselMbD/DirectionCosineConstraintIJ.h
@@ -17,7 +17,7 @@ namespace MbD {
     {
         //axisI axisJ aAijIeJe 
     public:
-        DirectionCosineConstraintIJ(EndFrmsptr frmi, EndFrmsptr frmj, int axisi, int axisj);
+        DirectionCosineConstraintIJ(EndFrmsptr frmi, EndFrmsptr frmj, size_t axisi, size_t axisj);
 
         void calcPostDynCorrectorIteration() override;
         virtual void initaAijIeJe();
@@ -36,7 +36,7 @@ namespace MbD {
         void preDynOutput() override;
 		void postDynOutput() override;
 
-        int axisI, axisJ;
+        size_t axisI, axisJ;
         std::shared_ptr<DirectionCosineIecJec> aAijIeJe;
     };
 }
